Split validation classes into their own files.
