+++ AutoRenter.Api.Validation/IValidatorFactory.cs
@@ -1,11 +1,11 @@
-﻿using System;
-using FluentValidation;
+﻿using FluentValidation;
 
 namespace AutoRenter.Api.Validation
 {
     public interface IValidatorFactory
     {
-        IValidator<T> GetValidator<T>();
-        IValidator GetValidator(Type type);
+        IValidator GetInsertInstance<T>();
+        IValidator GetUpdateInstance<T>();
+        IValidator GetDeleteInstance<T>();
     }
 }
