+++ AutoRenter.Api.Validation/LocationDeleteValidator.cs
@@ -1,0 +1,13 @@
+﻿using AutoRenter.Domain.Models;
+using FluentValidation;
+
+namespace AutoRenter.Api.Validation
+{
+    public class LocationDeleteValidator : AbstractValidator<Location>, IValidator<Location>
+    {
+        public LocationDeleteValidator()
+        {
+            RuleFor(m => m.Id).NotNull();
+        }
+    }
+}
