+++ AutoRenter.Api.Validation/VehicleDeleteValidator.cs
@@ -1,0 +1,13 @@
+﻿using AutoRenter.Domain.Models;
+using FluentValidation;
+
+namespace AutoRenter.Api.Validation
+{
+    public class VehicleDeleteValidator : AbstractValidator<Vehicle>, IValidator<Vehicle>
+    {
+        public VehicleDeleteValidator()
+        {
+            RuleFor(m => m.Id).NotNull();
+        }
+    }
+}
